Move AllowAnonymousAccessForDevelopmentHandler namespace
